Added validate_uuid4() method to utils.Refactoring.
